+++ lib/android_graphics.dart
@@ -6,6 +6,10 @@
 /// See: https://developer.android.com/reference/android/graphics/package-summary
 ///
 /// {@canonicalFor bitmap.Bitmap}
+/// {@canonicalFor point.Point}
+/// {@canonicalFor pointf.PointF}
 library android_graphics;
 
 export 'src/graphics/bitmap.dart' show Bitmap;
+export 'src/graphics/point.dart' show Point;
+export 'src/graphics/pointf.dart' show PointF;