+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.6] - 2019-02-13
+## [0.3.0] - 2019-02-14
 ### Added
 - `android_graphics.Bitmap.fromImage` static method
 - `android_graphics.Bitmap.fromAssetImage` static method
@@ -148,7 +148,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `android_content.Context.packageName` getter
 - `android_content.Context.packageResourcePath` getter
 
-[0.2.6]:  https://github.com/drydart/flutter_android/compare/0.2.5...0.2.6
+[0.3.0]:  https://github.com/drydart/flutter_android/compare/0.2.5...0.3.0
 [0.2.5]:  https://github.com/drydart/flutter_android/compare/0.2.4...0.2.5
 [0.2.4]:  https://github.com/drydart/flutter_android/compare/0.2.3...0.2.4
 [0.2.3]:  https://github.com/drydart/flutter_android/compare/0.2.2...0.2.3